Changed the method of getting a port's unique-code (identifier in the sub-modules of a composite module) from the port's pointer to be compatible with the composite-module architecture.
In the previous implementation where is not the composite-module architecture, each port has its own name.  Its unique-code can be simply defined as its host module's identifier + "." + its name (e.g. "environment.signal_start_of_episode).  But, in the current composite-module architecture, a port of a module of a composite module can be exported as a port of the composite module where the new port can have arbitrary name.  This exporting does not make a new port object, but just register the port's pointer with a new name.  Thus, to identify a port's unique-code from its pointer is not easy.  We need to write a search algorithm.  First, we have to identify the host module because several modules can have the same port pointer.  Then, we identify the port's identifier by search its pointer in the host module.

The main modification of this version is, therefore, to change the method of getting a port's unique-code from its pointer.  Important changes are: the TPortInterface class's member functions, Name, UniqueCode, and OuterBase are erased.  The PortUniqueCode member function is added to the TModuleInterface class.  The TCompositeModule class's ForEachSubConnection member function is changed (precise port information (TPortInfo class) is given to the input function object).  The parameter of the port templates' ConnectedPortFind member function is changed from a port's unique-code to a port's pointer.